remove user dialog

+++ src/app/users/users.component.ts
@@ -1,12 +1,12 @@
-import {Component, ElementRef, OnInit, ViewChild} from '@angular/core';
+import {Component, OnInit, ViewChild} from '@angular/core';
 import {select, Store} from '@ngrx/store';
 import * as userActions from './store/user.actions';
-import * as todoActions from '../todos/store/todo.actions';
 import * as fromUser from '../users/store/user.selectors';
 import {User} from './store/user.model';
 import {Observable} from 'rxjs';
-import {MDBModalRef, MDBModalService} from 'angular-bootstrap-md';
 import {DialogComponent} from '../dialog/dialog.component';
+import {DialogService} from '../dialog/dialog.service';
+import {MatDialog} from '@angular/material/dialog';
 
 @Component({
   selector: 'app-users',
@@ -18,7 +18,9 @@ export class UsersComponent implements OnInit {
   userList$: Observable<Array<User>>;
   @ViewChild('removeDialog', {static: false}) removeDialogComponent: DialogComponent;
 
-  constructor(private store: Store<any>, private modalService: MDBModalService) {
+  constructor(private store: Store<any>,
+              private dialogService: DialogService,
+              public dialog: MatDialog) {
   }
 
   ngOnInit() {
@@ -38,13 +40,18 @@ export class UsersComponent implements OnInit {
     this.editField = event.target.textContent;
   }
 
-  remove(user: User) {
-    this.removeDialogComponent.title = 'User Removal';
-    this.removeDialogComponent.content = 'Are you sure you want to remove the user: ' + user.name + ' ?';
-    this.removeDialogComponent.isDeleteType = true;
-    this.removeDialogComponent.open();
-    this.removeDialogComponent.approve$.subscribe(() => {
-      this.store.dispatch(userActions.deleteUser({id: user.id}));
+  openRemoveDialog(user: User): void {
+    const dialogRef = this.dialog.open(DialogComponent, {
+      width: '800px',
+      data: {title: 'User Removal', content: 'Are you sure you want to remove the user: ' + user.name + ' ?'},
+      panelClass: 'dialog-padding'
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      console.log('The dialog was closed');
+      if (!!result) {
+        this.store.dispatch(userActions.deleteUser({id: user.id}));
+      }
     });
   }
